+++ sandbox/sandbox.py
@@ -1,5 +1,7 @@
+from source.engines.daq_engine import DaqEngine
+from source.engines.logger_engine import LoggerEngine
+from source.engines.user_workflow_engine import UserWorkflowEngine
 from source.messaging import PubSubMessageCenter
-from source.engines import LoggerEngine, DaqEngine, UserWorkflowEngine
 from source.daq_drivers import SimulatedDaqDriver
 from source.data_logging_drivers import TdmsLoggerDriver, FileLoggingSession, JsonLoggerDriver
 from pathlib import Path
