--- source/data_types.py
@@ -1,7 +1,0 @@
-# import numpy as np
-#
-#
-# class DataPoint:
-#     """ I have no idea if this class is a good idea"""
-#     def __init__(self):
-#         self._
