+++ source/engines/_engines_shared.py
@@ -1,0 +1,9 @@
+from abc import abstractmethod
+
+DATA_TOPIC = "data"
+
+
+class _EngineMessage:
+    @abstractmethod
+    def execute(self, daq_engine):
+        pass
