+++ source/main.py
@@ -2,9 +2,3 @@ from .ui.main_window import start_app
 
 if __name__ == "__main__":
     start_app()
-    # thermocouple_device = Device()
-    # water_heater = Device()
-    # temperature_controller = ()
-    # while(1):
-    #
-    #     thermocouple_device.write_outputs()
